+++ Reversi/GameSampleState.h
@@ -10,7 +10,7 @@ public:
 	virtual void Init() override;
 	virtual void Update(float dt) override;
 	virtual void Render(float dt) override;
-	virtual void MouseInput(const sf::Vector2i& pos) override;
+	virtual void MouseInput(const sf::Vector2f& pos) override;
 	virtual void KeyInput(sf::Keyboard::Key key) override;
 };
