+++ Reversi/State.h
@@ -17,7 +17,7 @@ public:
 	virtual void Init() = 0;
 	virtual void Update(float dt) = 0;
 	virtual void Render(float dt) = 0;
-	virtual void MouseInput(const sf::Vector2i& pos) = 0;
+	virtual void MouseInput(const sf::Vector2f& pos) = 0;
 	virtual void KeyInput(sf::Keyboard::Key key) = 0;
 };
